% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_predict.R
\name{cv_predict}
\alias{cv_predict}
\alias{cv_predict.cv}
\alias{cv_predict.model}
\alias{cv_predict.multimodel}
\alias{cv_predict.default}
\alias{cv_resid}
\alias{cv_resid.cv}
\alias{cv_resid.model}
\alias{cv_resid.multimodel}
\alias{cv_resid.default}
\title{Extract out-of-sample predictions and residuals from cross-validation}
\usage{
cv_predict(x, ...)

\method{cv_predict}{cv}(x, simplify = TRUE, ...)

\method{cv_predict}{model}(x, simplify = TRUE, ...)

\method{cv_predict}{multimodel}(x, simplify = TRUE, ...)

\method{cv_predict}{default}(x, simplify = TRUE, ...)

cv_resid(x, ...)

\method{cv_resid}{cv}(x, simplify = TRUE, ...)

\method{cv_resid}{model}(x, ...)

\method{cv_resid}{multimodel}(x, simplify = TRUE, ...)

\method{cv_resid}{default}(x, simplify = TRUE, ...)
}
\arguments{
\item{x}{A \dQuote{cv} object or an object of another class. See \dQuote{Methods}.}

\item{\dots}{Not used in \code{*.cv} methods, passed to \code{cv} in all other methods.}

\item{simplify}{Logical: If \code{FALSE}, a list of vectors will be returned in case of multiple models,
if \code{TRUE} the output will be converted to a matrix.}
}
\value{
A vector, or if there is more than one model,
a matrix (if \code{simplify=TRUE}) or list of vectors (\code{simplify=FALSE}).
}
\description{
\code{cv_predict()} extracts out-of-sample predictions from a \dQuote{cv} or related object,
\code{cv_resid()} extracts out-of-sample residuals. Both are generic.
}
\details{
The output will depend on the \code{folds} used in \code{\link{cv}()}.
If the folds do not include all observations in the model data,
the output vector will be shorter than the number of rows in the model data.
}
\section{Methods}{

\itemize{
\item{\code{cv_predict.cv(x, ...)} is the core method described above.}
\item{\code{cv_predict.model(x, ...)} and \code{cv_predict.multimodel} execute
\code{x \%>\% cv(...) \%>\% cv_predict}.}
\item{\code{cv_predict.default(x, ...)} is applied to a fitted model; it executes \code{x \%>\% model \%>\% cv(...) \%>\% cv_predict}.}
}
\code{cv_resid} has the same methods as \code{cv_predict}.
}

\examples{
# Simulate data 
set.seed(123)
n <- 30
p <- 20
x <- matrix(rnorm(p*n), nrow = n)
y <- rowSums(x) + rnorm(n)
mymodel <- model(lm(y~x))

# Compare correlation of in-sample and out-of-sample predictions
# with response
preds <- data.frame(
  training = predict(mymodel), 
  test = drop(cv_predict(mymodel)))
cor(response(mymodel), preds)

}
\seealso{
\code{\link{response}}, \code{\link{predict.model}}
}
