% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_metric.R
\name{default_metric}
\alias{default_metric}
\title{Get the default metric of an object}
\usage{
default_metric(x)
}
\arguments{
\item{x}{A \link{model}, \link{multimodel}, fitted model or an object of class \dQuote{\link{cv}}.}
}
\value{
A named list of length one containing the metric function, if available.
}
\description{
This generic function returns the default metric of a model.
The default metric is the metric that is used when evaluating a model with
\code{\link{cv_performance}()}, \code{\link[modeltuner]{performance}()} or
\code{\link{evaluation_log}()} and no \code{metric} argument is specified.

The default metric of a \emph{model} is usually \code{\link[modeltuner]{rmse}} if the response is continuous and
\code{\link[modeltuner]{logLoss}} if it is binary.
The only exception to this rule are models based on \code{\link{fm_xgb}()},
where the \code{eval_metric} used in \code{\link[xgboost]{xgb.train}()} is chosen as default.

If \code{x} is a \link{multimodel} or \dQuote{cv} object with several models,
then the first model's default metric is returned.
}
\examples{
n <- 100
d <- data.frame(x = rnorm(n), y = rnorm(n))
lm(y~x, d) |> default_metric()
glm(I(y>0)~x, d, family = binomial) |> default_metric()
fm_xgb(y~x, d, nrounds = 10) |> default_metric()
fm_xgb(y~x, d, nrounds = 10, objective = "reg:absoluteerror") |> default_metric()
fm_xgb(y~x, d, nrounds = 10, objective = "reg:pseudohubererror") |> 
default_metric()  # -> there is no function with this name

}
\seealso{
\code{\link{metrics}}, \code{\link{fm_xgb}}
}
