% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_formula.R
\name{expand_formula}
\alias{expand_formula}
\title{Expand a formula}
\usage{
expand_formula(formula, ..., expand = TRUE)
}
\arguments{
\item{formula}{A formula.}

\item{\dots}{Named parameters to be expanded.}

\item{expand}{Logical: Expand the \dQuote{...} arguments (default) or join them element-wise?
If \code{expand=TRUE}, the vectors in \dQuote{...} will be expanded, the number of models will equal the product
of the lengths of the \dQuote{...} arguments;
otherwise, all \dQuote{...} arguments must have equal lengths, and the number of models will be equal to their common length.}
}
\value{
A list as \code{formula}s.
}
\description{
Replicate a formula with varying parameter values.
}
\examples{
expand_formula(y ~ ns(x, df = k), k = 1:10)
expand_formula(y ~ I(x^e1) + I(z^e2), e1 = 1:2, e2 = 1:3)
expand_formula(y ~ I(x^e1) + I(z^e2), e1 = 1:3, e2 = 1:3, expand = FALSE)

}
