% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fits.R
\name{extract_fits}
\alias{extract_fits}
\title{Extract the models fitted in a cross-validation from a \dQuote{cv} object.}
\usage{
extract_fits(x, ...)
}
\arguments{
\item{x}{A \code{cv}-object.}

\item{\dots}{Currently not used.}
}
\value{
A \link{param_table}.
Besides the model fits, it will also have a column containing the \code{folds}.
}
\description{
Extraction of the model fits is only possible if these have been saved as a part of the \dQuote{cv} object,
which is not the case by default. See the argument \code{keep_fits} in \code{?\link{cv}}.
}
\examples{
mm <- c(lm =    model(lm(Sepal.Width ~ ., iris)), 
        if (require(rpart)) model(rpart(Sepal.Width ~ ., iris), label = "rpart"))
mycv <- cv(mm, keep_fits = TRUE)
extract_fits(mycv)

}
\seealso{
\code{\link{cv}}
}
