% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R
\name{extract_model}
\alias{extract_model}
\alias{extract_model.multimodel}
\alias{extract_model.cv}
\alias{extract_multimodel}
\alias{extract_multimodel.cv}
\title{Extraction of a model and multimodel}
\usage{
extract_model(x, which, ...)

\method{extract_model}{multimodel}(x, which, use_cv_info = TRUE, ...)

\method{extract_model}{cv}(x, which, use_cv_info = TRUE, ...)

extract_multimodel(x, ...)

\method{extract_multimodel}{cv}(x, use_cv_info = TRUE, ...)
}
\arguments{
\item{x}{A \dQuote{multimodel} or \dQuote{cv} object.}

\item{which}{Selection of one model: An integer value
or a logical vector of length \code{n_model(x)} having exactly one \code{TRUE},
or a character value (selection by name, a model's \code{label}).
If \code{n_model(x)==1}, the only model is selected by default.}

\item{\dots}{Arguments passed from or to methods.}

\item{use_cv_info}{Logical: Whether to set the preferred iteration according to
results from the cross-validation (if these are present).
Relevant only for \emph{iteratively fitted models} (\link{ifm}).}
}
\value{
\code{extract_model()} returns a \code{\link{model}},
\code{extract_multimodel()} returns a \code{\link{multimodel}}.
}
\description{
\code{extract_model()} extracts a model from a \dQuote{multimodel} or \dQuote{cv} object.

\code{extract_multimodel()} extracts the multimodel from a \dQuote{cv} object.
}
\examples{
mm_swiss <- c(model(lm(Fertility ~ Education, swiss)),
              model(lm(Fertility ~ Education + I(Education^2) , swiss)))
cv_swiss <- cv(mm_swiss)

extract_model(mm_swiss, 1) # a model
subset(mm_swiss, 1)        # a multimodel

extract_model(cv_swiss, 1)      # a model
extract_multimodel(cv_swiss, 1) # a multimodel
subset(cv_swiss, 1)             # a cv

}
\seealso{
\code{\link{multimodel}}, \code{\link{cv}}
}
