% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_smooth_spline.R
\name{fm_smooth_spline}
\alias{fm_smooth_spline}
\alias{predict.fm_smooth_spline}
\title{Smoothing spline model with \code{formula}-interface}
\usage{
fm_smooth_spline(formula, data, weights = NULL, na.action = na.omit, ...)

\method{predict}{fm_smooth_spline}(object, newdata, ...)
}
\arguments{
\item{formula}{A \code{formula} with only one variable on right hand side.}

\item{data}{A \code{data.frame}.}

\item{weights}{Case \code{weights}.}

\item{na.action}{A function which indicates what should happen when the data contain \code{NA}s.
\code{\link{na.omit}} is the default, \code{\link{na.exclude}} or \code{\link{na.fail}}
could be useful alternative settings.}

\item{\dots}{Passed to \code{\link{smooth.spline}()}.}

\item{object}{Object of class \dQuote{fm_smooth_spline}.}

\item{newdata}{\code{data.frame} with the data to be predicted.
If missing, predictions for the model data will be returned.}
}
\value{
\code{fm_smooth_spline()} returns a list of class \dQuote{fm_smooth_spline} with components
\itemize{
\item{\emph{smooth.spline}: the fitted spline, of class \dQuote{\link{smooth.spline}};}
\item{\emph{formula}: the formula;}
\item{\emph{weights}: the fitting weights;}
\item{\emph{na.action}: the \code{\link{na.action}} used during data preparation;}
\item{\emph{call}: the matched call generating the model.}
}
}
\description{
This is a simple wrapper for \code{\link{smooth.spline}()} from the \code{stats} package.
}
\section{Methods (by generic)}{
\itemize{
\item \code{predict(fm_smooth_spline)}: \code{predict} method for class \dQuote{fm_smooth_spline}.

}}
\examples{
(smooth <- fm_smooth_spline(dist ~ speed, data = cars))
newd <- data.frame(speed = seq(min(cars$speed), max(cars$speed), length.out = 100))
plot(dist ~speed, cars, col = 4)
lines(newd$speed, predict(smooth, newd), col = 2)

}
\seealso{
\code{\link[stats]{smooth.spline}} (package \strong{stats})
}
