% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\name{label}
\alias{label}
\alias{label<-}
\alias{set_label}
\alias{n_model}
\title{Query or set model label(s)}
\usage{
label(x, which)

label(x, which) <- value

set_label(x, value, which = seq_along(label(x)))

n_model(x)
}
\arguments{
\item{x}{Object of class \dQuote{model}, \dQuote{multimodel}, \dQuote{cv} or \dQuote{performance}.}

\item{which}{Optional integer index  - allows querying or setting only a subset of the model labels.
Must have the same length as \code{value}. Default: All models in \code{x}.}

\item{value}{Labels to be attributed to models.}
}
\value{
\code{label()} returns a character vector.
\code{set_label()} returns its input object \code{x} with modified labels.
\code{n_model()} returns an integer.
}
\description{
\code{label()} extracts the model label(s) from \code{x}.
The replacement functions \verb{label<-} and \code{set_label()} are used to set model label(s) in \code{x}.
\code{n_model()} returns the number of models included in an object \code{x}.
}
\details{
These functions are generic and have methods for classes \dQuote{model}, \dQuote{multimodel}, \dQuote{cv},
\dQuote{performance}, \dQuote{evaluation_log}.

\code{n_model()} just returns \code{length(label(x))} and is thus available for any object having an appropriate \code{label()} method.
}
\examples{
xvars <- names(iris)[-1]
forms <- lapply(xvars, reformulate, response = names(iris)[1])
model0 <- model(lm(Sepal.Length~1, iris), label = "intercept_only")
mm <- c(model0, multimodel(model0, formula = forms))
n_model(mm)

label(mm)  
label(mm, 2:5) <- xvars
label(mm)
mm

set_label(mm, paste0("model_number_", 1:5))

}
\seealso{
\code{\link{model}}, \code{\link{multimodel}}, \code{\link{cv}}, \code{\link{cv_performance}},
\code{\link{evaluation_log}}
}
