% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_folds.R
\name{make_folds}
\alias{make_folds}
\title{Create folds (cross-validation groups)}
\usage{
make_folds(n, nfold = getOption("cv_nfold"), folds = NULL, strata = NULL)
}
\arguments{
\item{n}{An integer (corresponding to the number of observations in data) or a \code{data.frame}.}

\item{nfold}{Either an integer vector of length 1 or 2, or a numeric value between 0 and 1
(see \dQuote{Details}).}

\item{folds}{A list of integer vectors (optional), predetermined group structure.
If \code{folds} is given, the arguments \code{nfold} and \code{strata} will be ignored.}

\item{strata}{A vector or list of vectors: Strata for cross validation.
If specified and in accordance with \code{n}, the groups defined in \code{folds} are output,
ignoring \code{n} and \code{nfold}.}
}
\value{
A list of integer vectors defining the test sets, each of them a subset of \code{1:n} (or \code{1:nrow(n)} if \code{n} is a \code{data.frame}),
of class \dQuote{folds}.
}
\description{
Obtains test sets (\dQuote{folds}) for cross-validation procedures with \code{\link{cv}()}.
The user inputs specifications of the \emph{test} sets; the respective complements will be taken as \emph{training} sets.
}
\details{
There are three ways to define the number of groups and group sizes in \code{nfold}:
\itemize{
\item{A positive integer value: \emph{Complete} \code{nfold}-fold cross-validation;
\code{1:n} is partitioned into \code{nfold} groups of (nearly) equal size at random.
}
\item{Two positive integer values, the second smaller than the first: \emph{Incomplete} \code{nfold}-fold cross-validation;
\code{1:n} is partitioned into \code{nfold[1]} groups of (nearly) equal size at random, but only
\code{nfold[2]} of them are kept.
}
\item{A numeric value between 0 and 1: Hold-out validation;
there will be only one test group that will contain approximately \code{n*nfold} indices from \code{1:n}.
}
}
}
\examples{
make_folds(100, 10)       # Complete 10-fold CV
make_folds(100, c(10, 4)) # Incomplete 10-fold CV
make_folds(100, 0.3)      # Hold-out Validation with 30\% test data
make_folds(100, c(3, 1))  # Almost the same as make_folds(100, 1/3)
make_folds(iris)          # data as input
make_folds(100, folds = list(1:10, 11:40, 41:100))  # Unequal group sizes

}
\seealso{
\code{\link{cv}}
}
