% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics}
\alias{metrics}
\alias{rmse}
\alias{mae}
\alias{medae}
\alias{mse}
\alias{logLoss}
\alias{classification_error}
\title{Metrics}
\usage{
rmse(actual, predicted, w = NULL, ...)

mae(actual, predicted, w = NULL, ...)

medae(actual, predicted, w = NULL, ...)

mse(actual, predicted, w = NULL, ...)

logLoss(actual, predicted, w = NULL, ..., eps = .Machine$double.neg.eps)

classification_error(actual, predicted, w = NULL, ..., cut_value = 0.5)
}
\arguments{
\item{actual}{Observed values.}

\item{predicted}{Predicted values.}

\item{w}{Optional case weights.}

\item{...}{Passed to other functions or methods.}

\item{eps}{Adjustment value in \code{logLoss}.}

\item{cut_value}{Cut value for binary classification.}
}
\value{
All of these functions return a numeric value.
}
\description{
A metric is a function having arguments \code{actual}, \code{predicted} and optionally \code{w} (weights)
and returning a non-negative real value.

The following metrics taken from package \CRANpkg{MetricsWeighted} are included in \code{modeltuner}:
\itemize{
\item{Metrics for continuous response:
\code{\link[MetricsWeighted]{rmse}}, \code{\link[MetricsWeighted]{mae}}, \code{\link[MetricsWeighted]{medae}} and
\code{\link[MetricsWeighted]{mse}}.}
\item{Metrics for binary response:
\code{\link[MetricsWeighted]{logLoss}} and \code{\link[MetricsWeighted]{classification_error}}.
}}
}
\details{
The two metrics for binary response are slightly different from their \code{MetricsWeighted} counterparts:
\itemize{
\item{
\code{logLoss()}: Values of 0 and 1 in \code{predicted} are replaced by \code{eps} and \code{1-eps}, respectively,
before applying \code{MetricsWeighted::\link[MetricsWeighted]{logLoss}()}.
This prevents errors in case of prediction values of exactly 0 or 1.}
\item{
\code{classification_error()}: \code{predicted} is replaced by \code{ifelse(predicted>=cut_value, 1, 0)}
before applying \code{MetricsWeighted::\link[MetricsWeighted]{classification_error}()}.
}}
}
\examples{
data(mcycle, package = "MASS")
mod <- lm(accel ~ times, mcycle)
actual <- mcycle$accel
pred <- predict(mod)
rmse(actual, pred)
medae(actual, pred)

# performance() uses these metrics:
performance(mod)
performance(mod, metric = "medae")
performance(mod, metric = list(med_abs_err = medae))
}
\seealso{
\code{\link{default_metric}}
}
