% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-methods.R
\name{model-methods}
\alias{model-methods}
\alias{model.glm}
\alias{model.glmrob}
\alias{model.gam}
\alias{model.ranger}
\alias{model.merMod}
\alias{model.lmerMod}
\alias{model.glmerMod}
\title{Special methods of \code{model()}}
\usage{
\method{model}{glm}(
  x,
  ...,
  predict_function = function(object, ...) predict(object, ..., type = "response"),
  env = parent.frame()
)

\method{model}{glmrob}(
  x,
  ...,
  predict_function = function(object, ...) predict(object, ..., type = "response"),
  env = parent.frame()
)

\method{model}{gam}(
  x,
  ...,
  predict_function = function(object, ...) predict(object, ..., type = "response"),
  env = parent.frame()
)

\method{model}{ranger}(
  x,
  ...,
  predict_function = function(object, ...) predict(object, ...)$predictions,
  env = parent.frame()
)

\method{model}{merMod}(
  x,
  ...,
  predict_function = function(object, ..., type = "response", allow.new.levels = TRUE)
    predict(object, ..., allow.new.levels = allow.new.levels),
  env = parent.frame()
)

\method{model}{lmerMod}(
  x,
  ...,
  predict_function = function(object, ..., type = "response", allow.new.levels = TRUE)
    predict(object, ..., allow.new.levels = allow.new.levels),
  env = parent.frame()
)

\method{model}{glmerMod}(
  x,
  ...,
  predict_function = function(object, ..., type = "response", allow.new.levels = TRUE)
    predict(object, ..., allow.new.levels = allow.new.levels),
  env = parent.frame()
)
}
\arguments{
\item{x}{A fitted model.}

\item{\dots}{Passed to \code{model.default()}.}

\item{predict_function}{As in \code{\link{model.default}}.}

\item{env}{An environment. Used for internal purposes.}
}
\value{
All these methods of \code{model()} return a \link{model}.
}
\description{
These methods account for special features of the \code{predict} methods of some popular model types:
\code{\link[stats]{glm}} (from package \strong{stats}),
\code{\link[robustbase]{glmrob}} (from package \CRANpkg{robustbase}),
\code{\link[mgcv]{gam}} (from package \CRANpkg{mgcv}),
\code{\link[ranger]{ranger}} (from package \CRANpkg{ranger}),
\code{\link[lme4]{lmer}}, \code{\link[lme4]{glmer}} (from package \CRANpkg{lme4}).

They all execute \code{\link{model.default}()} with an adjusted default value of \code{predict_function}.
}
\examples{
# Simulate data
d <- simuldat()

# ranger fitted model (random forest):
if (require(ranger)){
  ranger_fitted <- ranger(Y ~ ., d)
  # Methods predict.ranger() returns a list:
  str(predict(ranger_fitted, data = d))
  
  # Method model.ranger() makes sure that a "model" object returns
  # a vector, as required:
  ranger_model <- model(ranger_fitted)
  predict(ranger_model)
}

}
