% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Combine several fitted models in a multimodel}
\usage{
models(..., .env = parent.frame())
}
\arguments{
\item{\dots}{Passed to \code{\link{c.model}()}, after application of
\code{\link{model}()} to \emph{fitted} models in the list of \dQuote{...} arguments.}

\item{.env}{For internal use.}
}
\value{
A \link{multimodel}.
}
\description{
The function \code{models()} is a more general version of \code{\link{c.model}()}.
While the latter expects arguments inheriting of classes \dQuote{model} or \dQuote{multimodel}
in its \dQuote{...} arguments, the former also accepts \emph{fitted} models (e.g. an \dQuote{lm}).
}
\examples{
mm <- models(
  lm = lm(Sepal.Length ~ ., iris),
  rpart = model(rpart::rpart(Sepal.Length ~ ., iris)),
  xgboost = tune(fm_xgb(Sepal.Length ~ ., iris))
)
mm
cv_performance(mm)
  
}
