% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cv_info.R, R/crit_iter.R, R/cv.R,
%   R/cv_predict.R, R/cv_simple.R, R/evaluation_log.R, R/extract_model.R,
%   R/fm_xgb_cv.R, R/get_xy.R, R/model.R, R/modeltuner-internal.R
\name{add_cv_info}
\alias{add_cv_info}
\alias{prnt_cv_info}
\alias{eval_crit}
\alias{has_cv_fits}
\alias{oos_indices}
\alias{cv_simple}
\alias{evlog_int}
\alias{extract}
\alias{cb_xgb_iteration}
\alias{get_xy}
\alias{get_xy2}
\alias{prnt_compact}
\alias{modeltuner-internal}
\title{Internal functions in package \code{modeltuner}}
\usage{
add_cv_info(x, cvobj, ...)

prnt_cv_info(x, ...)

eval_crit(crit, x, ...)

has_cv_fits(x)

oos_indices(folds)

cv_simple(x, folds, metric, iter, keep_fits, verbose, ...)

evlog_int(classObj, x, metric, eval_weights, na.rm, ...)

extract(x, i)

cb_xgb_iteration(period = 1, prefix = "", keep_final_msg = FALSE)

get_xy(object, data, weights = NULL, ...)

get_xy2(data, weights = NULL, ...)

prnt_compact(x, ...)
}
\description{
These functions are mainly for internal purposes and will normally not be called directly by the user.
}
\details{
\code{add_cv_info()} is a generic auxiliary function adding info from cv to a model. Apply to model, returns model.

\code{eval_crit()} is a generic auxiliary function evaluating preference criteria for a given \link{ifm}s.

\code{has_cv_fits()} is a generic auxiliary function that discerns whether a cv contains the
cross-validation fits or not.

\code{oos_indices()} identifies the out-of-sample indices in a \code{predictions} matrix.

\code{cv_simple()} is a generic function performing cross-validation for one model.

\code{evlog_int()} is a generic function performing internal calculations
in \code{\link{evaluation_log}()}.

\code{extract()} is a generic extraction function used internally.

\code{cb_xgb_iteration()} is a callback function used when a xgboost model is cross-validated
if \code{verbose=TRUE}.
It is a reduced version of \code{\link[xgboost]{cb.print.evaluation}}

\code{get_xy()} is a generic taking formula and data, returns list of x matrix, y vector and terms object.
Designed for internal use in \code{formula}-interface wrappers.

\code{get_xy2()} is a generic dispatching on the second argument, \code{data}.

\code{prnt_compact()} and \code{prnt_cv_info()} are generic utility functions used in some \code{print} methods.
}
\keyword{internal}
