% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltuner_cheatsheet.R
\name{modeltuner_cheatsheet}
\alias{modeltuner_cheatsheet}
\title{Open a modeltuner cheatsheet}
\usage{
modeltuner_cheatsheet(open = TRUE)
}
\arguments{
\item{open}{Logical: If \code{TRUE}, a cheatsheet in pdf format is opened (via shell command);
If \code{FALSE}, the path to the pdf file is returned as a character string.}
}
\value{
Path to the cheatsheet in pdf format
}
\description{
\code{modeltuner_cheatsheet()} opens a pdf file (\dQuote{cheatsheet}) summarizing
important concepts and functions in the \code{modeltuner} package.
}
\examples{
\dontrun{modeltuner_cheatsheet()}

# Display path to pdf:
modeltuner_cheatsheet(open = FALSE)

}
