% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{modeltuner_options}
\alias{modeltuner_options}
\title{List all options defined in package \dQuote{modeltuner}}
\usage{
modeltuner_options()
}
\value{
Returns a list of option settings.
}
\description{
\code{modeltuner_options()} displays the current values of all options set in the package.
These options define default values for certain function parameters.
}
\details{
\itemize{
\item{\strong{cv_nfold} (value at package startup: \code{=10}): Default value of argument \code{nfold} in \code{\link{cv}()}, \code{\link{tune}()} and
\link{stepwise} functions.}
\item{\strong{cv_verbose}  (\code{=\link{interactive}()} at startup): Default value of argument \code{verbose} in \code{\link{cv}()},}
\item{\strong{cv_show_se} (\code{=FALSE} at startup): Default value of argument \code{se} in \code{\link{print.performance}()} and
\code{\link{print.evaluation_log}()} and default of \code{errorbars} in \code{\link{plot.evaluation_log}()},}
\item{\strong{cv_iter} (\code{=\link{crit_min}()} at startup): Default value of argument \code{iter} in \code{\link{cv}()} applied to models of class
\dQuote{\link{fm_xgb}} or \dQuote{\link{fm_glmnet}},}
\item{\strong{expand_max_model}  (\code{=30} at startup): Default value of argument \code{max_n_model} in
\code{\link{multimodel.model}()} and \code{\link{tune.model}()},}
\item{\strong{print_max_model} (\code{=3} at startup): Default value of \code{n} in \code{\link{print.multimodel}()}, \code{\link{print.cv}()} and
\code{\link{print.evaluation_log}()},}
\item{\strong{print_max_row} (\code{=30} at startup): Default value of \code{n} in \code{\link{print.performance}()} and
\code{print.\link{extract_pref_iter}()}},
\item{\strong{print_rows_evaluation_log} (\code{=6} at startup): Default value of \code{n_row} in \code{\link{print.evaluation_log}()}.}
}
}
\examples{
modeltuner_options()

}
