% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_aux.R
\name{null_formula}
\alias{null_formula}
\title{\dQuote{Null formula} of a model}
\usage{
null_formula(x)
}
\arguments{
\item{x}{Any object having a \code{\link{formula}} method and an \code{\link{update}} method
accepting updating formula as second argument.}
}
\value{
Returns a simple (\dQuote{null}) formula.
}
\description{
Returns the formula of an intercept-only model (\code{. ~ 1}) if the original formula \code{x} contains an intercept
and the constant zero-model formula (\code{. ~ 0}) otherwise.
}
\examples{
null_formula(lm(Sepal.Length ~ ., iris))
}
\seealso{
Used as default value of argument \dQuote{formula1} in \link{stepwise} functions.
}
