% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmodel.R
\name{pmodel}
\alias{pmodel}
\title{Purely predictive (non-fittable) model}
\usage{
pmodel(
  x,
  data,
  response,
  label = "pmodel",
  class = attr(x, "class")[[1]],
  weights = NULL,
  predict_function = predict
)
}
\arguments{
\item{x}{A fitted model.}

\item{data}{A \code{data.frame}.}

\item{response}{A character string, the name of the response.}

\item{label}{Character string: A label to be attributed to the output.}

\item{class}{Class of the model \code{x}.}

\item{weights}{Evaluation weights.}

\item{predict_function}{Predict function, often \code{\link{predict}}
(see \dQuote{Details} section in \code{?\link{model}}).}
}
\value{
An object of class \dQuote{pmodel} (also inheriting of class \dQuote{model}),
a simplified version of a \code{model}, including the components
\emph{label}, \emph{data}, \emph{response_type}, \emph{response}, \emph{fit},
\emph{predict_function}, \emph{class} and \emph{weights}.
Unlike a \code{model}, it does not have components \emph{formula} and \emph{call}
}
\description{
A \code{pmodel} is a \code{\link{model}} with reduced capacities.
It can be used to make predictions, but it cannot be updated or re-\link{fit}ted,
and therefore not cross-validated.
}
\details{
The model object \code{x} does not have to fulfill the formal requirements of a \link{model}
(model formula, call component, etc.).
The only requirement is that for any data set of the same shape as \code{data},
predictions can be obtained with the \code{predict_function}.

For tasks including only predicting, not re-fitting, the methods for class \code{model}
also work for a \code{pmodel}.
In particular, the following methods remain applicable:
\code{\link{predict.model}},
\code{\link{residuals.model}},
\code{\link{response.model}},
\code{\link{performance.model}},
\code{\link{plot.model}},
\code{\link{label.model}},
\code{\link{weights.model}}.
Among the generics with no appropriate method for a \code{pmodel} are
\code{\link{fit}} (with arguments inducing a modification of the model),
\code{\link{cv}} and all functions \code{cv_*}
(\code{\link{cv_performance}}, \code{\link{cv_predict}}, etc.).
}
\examples{
mylm <- lm(Sepal.Length ~ ., iris)
mypm <- pmodel(mylm, iris, "Sepal.Length")
mypm
predict(mypm)
residuals(mypm)
response(mypm)
performance(mypm)       
 
}
