% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.model.R
\name{predict.model}
\alias{predict.model}
\alias{residuals.model}
\alias{predict.multimodel}
\alias{residuals.multimodel}
\title{Predictions and residuals from a (multi-)model}
\usage{
\method{predict}{model}(object, newdata = object$data, ...)

\method{residuals}{model}(object, ...)

\method{predict}{multimodel}(object, newdata, ...)

\method{residuals}{multimodel}(object, ...)
}
\arguments{
\item{object}{A \code{\link{model}} or \code{\link{multimodel}}.}

\item{newdata}{newdata}

\item{\dots}{Passed to \code{\link{fit}}.}
}
\value{
A vector if \code{object} is a \dQuote{model}, a matrix if it is a \dQuote{multimodel} or \dQuote{cv} containing several models.
}
\description{
\code{predict.model()} and \code{residuals.model()} \code{\link{fit}} a model and
extract the \emph{in-sample} predictions and residuals, respectively.
}
\examples{
mod <- model(lm(Sepal.Length ~ .,  iris), 
             label = "lm_iris")
predict(mod)
residuals(mod)

}
\seealso{
\code{\link{fit}}, \code{\link{cv_predict}} (in particular examples), \code{\link{response}}.
}
