% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.R
\name{response}
\alias{response}
\alias{response.model}
\alias{response.default}
\alias{response.multimodel}
\alias{response.cv}
\title{Extract the values of the model response from an object}
\usage{
response(object, ...)

\method{response}{model}(object, ...)

\method{response}{default}(object, ...)

\method{response}{multimodel}(object, ...)

\method{response}{cv}(object, complete = FALSE, ...)
}
\arguments{
\item{object}{An object of class \dQuote{model}, \dQuote{multimodel} or \dQuote{cv} or a fitted model.}

\item{\dots}{Currently not used.}

\item{complete}{Logical: If \code{FALSE}, the output includes only the observations featuring exactly once in the
\code{folds} (sorted by their position in the model data); otherwise, it returns the full vector.}
}
\value{
A vector containing the values of the model^s response variable.
}
\description{
Extract the values of the model response from an object of class
\dQuote{\link{model}}, \dQuote{\link{multimodel}}, \dQuote{\link{cv}} or a fitted model.
}
\details{
The default method applies \code{response.model()} to \code{model(object, ...)}.

The method for class \dQuote{cv} with \code{complete=FALSE} returns a vector of the same length as
\code{\link{cv_predict.cv}(x)} and \code{\link{cv_resid.cv}(x)}.
}
\examples{
response(lm(mpg ~., mtcars))

}
\seealso{
\code{\link{cv_predict}} (in particular, the examples), \code{\link{predict.model}}.
}
