% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_iter.R
\name{set_iter}
\alias{set_iter}
\alias{set_iter.cv}
\alias{extract_iter}
\alias{expand_iter}
\alias{set_iter.model}
\alias{set_iter.model_fm_xgb}
\alias{set_iter.model_fm_glmnet}
\title{Set, extract and expand preference criteria for a \dQuote{cv} object for iteratively fitted models}
\usage{
set_iter(x, iter, ...)

\method{set_iter}{cv}(
  x,
  iter,
  which = label.cv(x),
  label = label.cv(x),
  keep_all = TRUE,
  ...
)

extract_iter(x, ...)

expand_iter(x, iter = NULL, which = label.cv(x))

\method{set_iter}{model}(x, ...)

\method{set_iter}{model_fm_xgb}(x, iter, verbose = TRUE, warn = TRUE, ...)

\method{set_iter}{model_fm_glmnet}(x, iter, lambda = NULL, verbose = TRUE, warn = TRUE, ...)
}
\arguments{
\item{x}{A cross-validated model, of class \dQuote{\link{cv}}, based on an \emph{iteratively fitted model} (IFM, see \link{ifm}).
In \code{set_iter()}, \code{x} can also be a \link{model} being based on an IFM.}

\item{iter}{If \code{x} is a \dQuote{cv}: A criterion (see \code{\link{crit_iter}}, multiple criteria not allowed).\cr
If \code{x} is a \dQuote{model}: An integer value; if model has cross-validation information, \code{iter} is not required and
the preferred iteration will be selected; otherwise \code{iter} is required.}

\item{\dots}{Arguments passed to methods.}

\item{which}{Character, integer or logical vector specifying the cross-validated models to be modified or expanded.}

\item{label}{\code{\link{label}} of output object.}

\item{keep_all}{Logical: If \code{TRUE}, all preference criteria are kept, if \code{FALSE}, only the selected one
(only relevant if there is more than one preference criterion).}

\item{verbose}{Logical: Show information on modified arguments?}

\item{warn}{Logical: Whether to issue a warning when information for parameter setting is not possible.}

\item{lambda}{Vector lambda (of decreasing numeric values) to pass to \code{\link{glmnet}()}.}
}
\value{
A modified object of the same class as \code{x}, a \dQuote{model} or \dQuote{cv} object.
}
\description{
\code{set_iter()} modifies a \dQuote{cv} object by altering its preferences with respect to iterations,
without re-running the cross-validation.

\code{extract_iter()} extracts the information on preferred iterations from a \dQuote{cv} object.

\code{expand_iter()} converts a \dQuote{cv} object with multiple preferred iterations to a \dQuote{cv} object
having several (identical) models, but different preferred iterations.
}
\examples{
d <- simuldat(n = 5000)
m <- model("fm_xgb", Y ~ ., d, nrounds = 200, class = "fm_xgb", 
           label = "xgb")
cvm <- cv(m, iter = c(crit_min(), crit_last(), crit_overfit(.5)), 
          nfold = .3)
print(cvm, what = "call")           
extract_iter(cvm)
cv_performance(cvm)
plot(evaluation_log(cvm))

# set_iter
cvm_last <- set_iter(cvm, crit_last(), label = "xgb_last")
cv_performance(c(cvm, cvm_last))

# expand_iter
cvm_expanded <- expand_iter(cvm)
print(cvm_expanded, what = c("call"))
cv_performance(cvm_expanded)       
plot(evaluation_log(cvm_expanded))

}
\seealso{
\code{\link{ifm}}, \code{\link{fm_xgb}}, \code{\link{fm_glmnet}}, \code{\link{fit}},
\code{\link{crit_iter}}.
}
