% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simuldat.R
\name{simuldat}
\alias{simuldat}
\title{Simulate data}
\usage{
simuldat(n = 500, ngroup = 20, df = Inf, b = 1, weights = NULL)
}
\arguments{
\item{n}{Number of observations}

\item{ngroup}{Number of groups in \code{g}.}

\item{df}{Number of degrees of freedom of t distribution for simulation of residuals.
Default \code{Inf} generates normal residuals.}

\item{b}{Numeric vector of length 4 (shorter will be expanded to length 4):
multiplicators of the 4 effects,}

\item{weights}{\code{NULL} or vector of \code{n} non-negative values:
If not \code{NULL}, simulated residuals will be divided by \code{1/sqrt(weights)}.}
}
\value{
\code{simuldat()} returns a \code{data.frame}.
}
\description{
Simulates a \code{data.frame} with \code{n} rows and 12 columns:
\itemize{
\item{\verb{X1, ..., X10}, random standard normal vectors.}
\item{\code{g}, a factor with \code{ngroup} levels, assigned at random.}
\item{\code{Y}, the response, generated as a additive combination of
\enumerate{
\item a linear effect in \code{X1} with (randomly generated) intercepts and slopes varying between groups of \code{g},
\item linear effects \code{X1}, \code{X2} and \code{X3}.
\item non-linear effects in \code{X6} and \code{X7}.
\item a non-linear three-way interaction of \code{X3}, \code{X4} and \code{X5} and
\item residuals from a normal or t distributions with \code{df} degrees of freedom (\code{df=Inf} is normal).}
}
See examples and code for details.
}
}
\examples{
# Simluate data
d <- simuldat()

# Fit the "true" model
if (require(lme4)){
  form <- Y ~ (X1|g) + X2 + X3 + abs(X3+X4+X5) + I(exp(2*X6)/(1+exp(2*X6))) + dnorm(X7, sd = 0.5)
  (truemodel <- lmer(form, d))
  \donttest{confint(truemodel)}
}

# Data set with a purely random response
str(simuldat(b = 0))

}
