% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{weights.model}
\alias{weights.model}
\title{Extract the (fitting) \code{weights} from a \dQuote{\link{model}} object}
\usage{
\method{weights}{model}(object, ...)
}
\arguments{
\item{object}{A \dQuote{model} object.}

\item{\dots}{Currently not used.}
}
\value{
Returns either a vector of weights or \code{NULL}.
}
\description{
Extract the (fitting) \code{weights} from a \dQuote{\link{model}} object
}
\examples{
# Simulate data
set.seed(314159)
n <- 50
x <- rnorm(n)
w <- runif(n)
y <- x + rnorm(n)/sqrt(w)

# Weighted lm
mymodel <- model(lm(y~x, weights = w))

# Extract weights
weights(mymodel)

}
\seealso{
\code{\link{weights}}, \code{\link{model}}
}
